projj/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
aiofiles>=23.0.0
playwright>=1.44.0
python-telegram-bot>=21.0.0
ollama>=0.2.0
apscheduler>=3.10.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
websockets>=12.0
pypdf>=4.0.0
python-docx>=1.1.0
fake-useragent>=1.5.0
